012_audio-transcriber/requirements.txt
assemblyai
python-dotenv
google-genai
pydantic-ai
pydantic-settings
requests
